Paginated Actions and Embeds for Reminder Edit View
# Overview
- The editing feature shows an embed with navigation buttons per page, and there are 5 embed fields per page, each representing a reminder item. A user can click on the button that matches up with field index to bring up a modal that allows the user to edit that entry.
- The edit feature was implemented by using the PaginatedMessage utility library from SapphireJs.
	- There were no known or existing solutions to implementing a paginated page with a unique button for each field entry.
	- Had to implement workaround by setting the interaction replied property to true after showing the modal when implementing the button.
- Models were created under models/pagination representing the data required for a paginated class.
- There is a static constructor method at the top level which uses multiple sub static constructors to ultimately convert an array of reminder data from the prisma DB to:
	- An array (ReminderPages) with Collections (ReminderPage) of field data + button action (ReminderComponentData) associated with it.
		- Each ReminderComponentData contains the data required for an EmbedBuilder to properly form and a button action built as a part of the static method pipelines.
			- The buttons are built from two ActionRows each containing one TextInputBuilder that are placed into a Modal.
			- The reminder ID is used as metadata for identifying them later in the modal interaction handler.
			- The customId is generated by serializing JSON data containing the reminder ID and type of interaction ("Reminder").
- The data types are not yet converted in case any further manipulation on the data is required.
- There are public methods to facilitate building the data required to show a component, see ReminderPaginatedResponseBuilder.generatePages() for a high level overview of how the general construction works.
- After the modal is submitted, a modal interaction handler uses a Zod pipeline to an incoming interaction's customId into reminder data (or reject it if it doesn't pass).
- The data is then retrieved from the modal and converted in usual fashion using the user's timezone data stored in the database, before storing the edited data.

# Patches
- Removed extraneous dependencies
# TODO
- Fix timezones being offset by wrong amount
	- Passing in "3h" should increase the time by 3h, but only increases it by 1h.
	- Passing in "1h" actually decreases the reminder time by 1h.
	- Likely there is something wrong during conversion process.
		- Will check and test functionality with the time string parser function (timestringToDayJsObj)
- Create tests
- Clean up logs in overrides of reminder builder.
- Handle user input if it is invalid.
- Add delete functionality

+++ src/index.ts
@@ -1,36 +1,36 @@
 import "./lib/setup";
 
-import { LogLevel, SapphireClient } from "@sapphire/framework";
-import { GatewayIntentBits, Partials } from "discord.js";
+import { LogLevel, SapphireClient } from "@sapphire/framework"
+import { GatewayIntentBits, Partials } from "discord.js"
 
-import { container } from "@sapphire/pieces";
-import prisma from "./db/prismaInstance";
-import { CooldownService } from "./features/cooldowns";
-import { TranslationService } from "./features/translation";
-import { logger } from "./logger/logger";
-import { ChannelService } from "./services/channelService";
-import { GuildService } from "./services/guildService";
-import { WebhookService } from "./services/webhookService";
+import { container } from "@sapphire/pieces"
+import prisma from "./db/prismaInstance"
+import { CooldownService } from "./features/cooldowns"
+import { TranslationService } from "./features/translation"
+import { logger } from "./logger/logger"
+import { ChannelService } from "./services/channelService"
+import { GuildService } from "./services/guildService"
+import { WebhookService } from "./services/webhookService"
 
 declare module "@sapphire/pieces" {
 	export interface Container {
-		translationService: TranslationService;
-		cooldownService: CooldownService;
+		translationService: TranslationService
+		cooldownService: CooldownService
 
-		prisma: typeof prisma;
-		webhookService: WebhookService;
-		dbLogger: typeof logger;
-		channelService: ChannelService;
-		guildService: GuildService;
+		prisma: typeof prisma
+		webhookService: WebhookService
+		dbLogger: typeof logger
+		channelService: ChannelService
+		guildService: GuildService
 	}
 }
 
-container.translationService = new TranslationService();
-container.cooldownService = new CooldownService();
+container.translationService = new TranslationService()
+container.cooldownService = new CooldownService()
 
-container.prisma = prisma;
-container.webhookService = new WebhookService();
-container.dbLogger = logger;
+container.prisma = prisma
+container.webhookService = new WebhookService()
+container.dbLogger = logger
 const client = new SapphireClient({
 	defaultPrefix: "!",
 	regexPrefix: /^(hey +)?bot[,! ]/i,
@@ -56,18 +56,22 @@ const client = new SapphireClient({
 	defaultCooldown: {
 		delay: 1_000,
 	},
-});
+})
 
 const main = async () => {
 	try {
-		client.logger.info("Logging in");
-		await client.login();
-		client.logger.info("logged in");
+		client.logger.info("Logging in")
+		await client.login()
+		client.logger.info("logged in")
 	} catch (error) {
-		client.logger.fatal(error);
-		client.destroy();
-		process.exit(1);
+		if (!client.isReady()) {
+			client.logger.fatal(error)
+			client.destroy()
+			process.exit(1)
+		} else {
+			client.logger.error(error)
+		}
 	}
-};
+}
 
 main();
